requirements.txt
openai
click
requests
beautifulsoup4
python-dotenv
newspaper3k
lxml[html_clean]
nltk
supabase
progress-parallel